Going with yp.@int, yp.str, and similar "constructors" (instead of "new yp_int", "new yp_str", etc)
This is fewer keystrokes (although the '@' is annoying, it's shorter than "new ") and avoids the confusion of creating a yp_list in C# only to have nohtyP transmute it underneath to a yp_list.  On the flip side, C# doesn't provide a way to shorten this further...the "yp." will always have to preceed the function name.