Disable /RTCc
These check for data truncation, but the error ignores explicit casts, requiring explicit bit masking instead. This is fine where the size of the types are known, but tedious for casts involving yp_ssize_t. Plus, we have type checking in other compilers to ensure the casts are at least explicit, and it's better to catch these errors in the build rather than in the tests. So disable this altogether.