Change the set/dict methods yp_unionN/etc to yp_union/etc
yp_unionN/etc were the only vararg functions that didn't take "unpacked" arguments. yp_tupleN takes the items to construct the tuple; yp_unionN should do the same. This is all the more obvious when you compare yp_updateK, which takes the items, to yp_updateN, which took the iterables. Should there ever be a yp_unionN in the future, the varargs should be interpreted as the items to add to the set/dict.